Restructure CV Base and Rewrite Main
