+++ compvis/Current Iteration/camera_calibration.py
@@ -1,0 +1,216 @@
+import cv2
+import numpy as np 
+import glob 
+import pickle
+
+def camera_calibration():
+    # TODO: Aiden -> implement camera calibration and store in new object
+    # Since this should only need to be completed one time per camera. Have the output saved as a .npy file with designation CAM1 and CAM2. 
+    # We will need to put a label on the cameras as different computers use different paths to the webcams (Unless someone can figure this out. Look into how CV uses indecies as an input to VideoStream function).
+    # Deciding which camera gets what matrix will happen by seeing what camera index gets used for the different cams and applying the correction Matricies to that camera.
+    # See video for more information: https://www.youtube.com/watch?v=uKDAVcSaNZA
+    ################ FIND CHESSBOARD CORNERS - OBJECT POINTS AND IMAGE POINTS #############################
+
+    chessboardSize = (8,5)
+    frameSize = (640,480)
+
+
+
+    # termination criteria
+    criteria = (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 30, 0.001)
+
+
+    # prepare object points, like (0,0,0), (1,0,0), (2,0,0) ....,(6,5,0)
+    objp = np.zeros((chessboardSize[0] * chessboardSize[1], 3), np.float32)
+    objp[:,:2] = np.mgrid[0:chessboardSize[0],0:chessboardSize[1]].T.reshape(-1,2)
+
+
+    # Arrays to store object points and image points from all the images.
+    objpoints = [] # 3d point in real world space
+    imgpointsL = [] # 2d points in image plane.
+    imgpointsR = [] # 2d points in image plane.
+
+
+    imagesLeft = glob.glob('images/stereoLeft/*.png')
+    imagesRight = glob.glob('images/stereoRight/*.png')
+
+    for imgLeft, imgRight in zip(imagesLeft, imagesRight):
+
+        imgL = cv2.imread(imgLeft)
+        imgR = cv2.imread(imgRight)
+        grayL = cv2.cvtColor(imgL, cv2.COLOR_BGR2GRAY)
+        grayR = cv2.cvtColor(imgR, cv2.COLOR_BGR2GRAY)
+
+        # Find the chess board corners
+        retL, cornersL = cv2.findChessboardCorners(grayL, chessboardSize, None)
+        retR, cornersR = cv2.findChessboardCorners(grayR, chessboardSize, None)
+
+        # If found, add object points, image points (after refining them)
+        if retL and retR == True:
+
+            objpoints.append(objp)
+
+            cornersL = cv2.cornerSubPix(grayL, cornersL, (11,11), (-1,-1), criteria)
+            imgpointsL.append(cornersL)
+
+            cornersR = cv2.cornerSubPix(grayR, cornersR, (11,11), (-1,-1), criteria)
+            imgpointsR.append(cornersR)
+
+            # Draw and display the corners
+            cv2.drawChessboardCorners(imgL, chessboardSize, cornersL, retL)
+            cv2.imshow('img left', imgL)
+            cv2.drawChessboardCorners(imgR, chessboardSize, cornersR, retR)
+            cv2.imshow('img right', imgR)
+            cv2.waitKey(1000)
+
+
+    cv2.destroyAllWindows()
+
+
+
+
+    ############## CALIBRATION #######################################################
+
+    retL, cameraMatrixL, distL, rvecsL, tvecsL = cv2.calibrateCamera(objpoints, imgpointsL, frameSize, None, None)
+    heightL, widthL, channelsL = imgL.shape
+    newCameraMatrixL, roi_L = cv2.getOptimalNewCameraMatrix(cameraMatrixL, distL, (widthL, heightL), 1, (widthL, heightL))
+
+    retR, cameraMatrixR, distR, rvecsR, tvecsR = cv2.calibrateCamera(objpoints, imgpointsR, frameSize, None, None)
+    heightR, widthR, channelsR = imgR.shape
+    newCameraMatrixR, roi_R = cv2.getOptimalNewCameraMatrix(cameraMatrixR, distR, (widthR, heightR), 1, (widthR, heightR))
+
+
+
+    ########## Stereo Vision Calibration #############################################
+
+    flags = 0
+    flags |= cv2.CALIB_FIX_INTRINSIC
+    # Here we fix the intrinsic camara matrixes so that only Rot, Trns, Emat and Fmat are calculated.
+    # Hence intrinsic parameters are the same 
+
+    criteria_stereo= (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 30, 0.001)
+
+    # This step is performed to transformation between the two cameras and calculate Essential and Fundamenatl matrix
+    retStereo, newCameraMatrixL, distL, newCameraMatrixR, distR, rot, trans, essentialMatrix, fundamentalMatrix = cv2.stereoCalibrate(objpoints, imgpointsL, imgpointsR, newCameraMatrixL, distL, newCameraMatrixR, distR, grayL.shape[::-1], criteria_stereo, flags)
+
+    #print(newCameraMatrixL)
+    #print(newCameraMatrixR)
+
+    ########## Stereo Rectification #################################################
+
+    rectifyScale= 1
+    rectL, rectR, projMatrixL, projMatrixR, Q, roi_L, roi_R= cv2.stereoRectify(newCameraMatrixL, distL, newCameraMatrixR, distR, grayL.shape[::-1], rot, trans, rectifyScale,(0,0))
+
+    stereoMapL = cv2.initUndistortRectifyMap(newCameraMatrixL, distL, rectL, projMatrixL, grayL.shape[::-1], cv2.CV_16SC2)
+    stereoMapR = cv2.initUndistortRectifyMap(newCameraMatrixR, distR, rectR, projMatrixR, grayR.shape[::-1], cv2.CV_16SC2)
+
+    print("Saving parameters!")
+    cv_file = cv2.FileStorage('stereoMap.xml', cv2.FILE_STORAGE_WRITE)
+
+    cv_file.write('stereoMapL_x',stereoMapL[0])
+    cv_file.write('stereoMapL_y',stereoMapL[1])
+    cv_file.write('stereoMapR_x',stereoMapR[0])
+    cv_file.write('stereoMapR_y',stereoMapR[1])
+
+    cv_file.release()
+
+
+    # termination criteria
+    criteria = (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 30, 0.001)
+
+
+    # prepare object points, like (0,0,0), (1,0,0), (2,0,0) ....,(6,5,0)    
+    objp = np.zeros((chessboardSize[0] * chessboardSize[1], 3), np.float32)
+    objp[:,:2] = np.mgrid[0:chessboardSize[0],0:chessboardSize[1]].T.reshape(-1,2)
+
+    size_of_chessboard_squares_mm = 20
+    objp = objp * size_of_chessboard_squares_mm
+
+
+    # Arrays to store object points and image points from all the images.
+    objpoints = [] # 3d point in real world space
+    imgpoints = [] # 2d points in image plane.
+
+
+    images = glob.glob('images/*.png')
+
+
+    for image in images:
+
+        img = cv2.imread(image)
+        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+
+        # Find the chess board corners
+        ret, corners = cv2.findChessboardCornersSB(gray, chessboardSize, flags=cv2.CALIB_CB_EXHAUSTIVE)
+        
+        # If found, add object points, image points (after refining them)
+        if ret == True:
+
+            objpoints.append(objp)
+            corners2 = cv2.cornerSubPix(gray, corners, (11,11), (-1,-1), criteria)
+            imgpoints.append(corners)
+
+            # Draw and display the corners
+            cv2.drawChessboardCorners(img, chessboardSize, corners2, ret)
+            cv2.imshow('img', img)
+            cv2.waitKey(1000)
+
+
+    cv2.destroyAllWindows()
+
+
+
+
+    ############## CALIBRATION #######################################################
+
+    ret, cameraMatrix, dist, rvecs, tvecs = cv2.calibrateCamera(objpoints, imgpoints, frameSize, None, None)
+
+    # Save the camera calibration result for later use (we won't worry about rvecs / tvecs)
+    pickle.dump((cameraMatrix, dist), open( "calibration.pkl", "wb" ))
+    pickle.dump(cameraMatrix, open( "cameraMatrix.pkl", "wb" ))
+    pickle.dump(dist, open( "dist.pkl", "wb" ))
+
+
+    # ############## UNDISTORTION #####################################################
+    count = 1
+    for image in images:
+        img = cv2.imread(image)
+        h,  w = img.shape[:2]
+        newCameraMatrix, roi = cv2.getOptimalNewCameraMatrix(cameraMatrix, dist, (w,h), 1, (w,h))
+
+
+
+        # Undistort
+        dst = cv2.undistort(img, cameraMatrix, dist, None, newCameraMatrix)
+
+        # crop the image
+        x, y, w, h = roi
+        dst = dst[y:y+h, x:x+w]
+        cv2.imwrite(f'res_image{count}.png', dst)
+
+
+        # Undistort with Remapping
+        mapx, mapy = cv2.initUndistortRectifyMap(cameraMatrix, dist, None, newCameraMatrix, (w,h), 5)
+        dst = cv2.remap(img, mapx, mapy, cv2.INTER_LINEAR)
+
+        # crop the image
+        x, y, w, h = roi
+        dst = dst[y:y+h, x:x+w]
+        cv2.imwrite(f'res_image{count}2.png', dst)
+        count += 1
+
+
+
+
+    # Reprojection Error
+    mean_error = 0
+
+    for i in range(len(objpoints)):
+        imgpoints2, _ = cv2.projectPoints(objpoints[i], rvecs[i], tvecs[i], cameraMatrix, dist)
+        error = cv2.norm(imgpoints[i], imgpoints2, cv2.NORM_L2)/len(imgpoints2)
+        mean_error += error
+
+    print( "total error: {}".format(mean_error/len(objpoints)) )
+
+
+    pass
