+++ compvis/Current Iteration/color_finder.py
