+++ compvis/Prev Files/color_findergpt.py
