+++ compvis/Prev Files/colordetec.py
@@ -39,7 +39,7 @@ def colordetec():
 	# if a video path was not supplied, grab the reference
 	# to the webcam
 	if not args.get("video", False):
-		vs = VideoStream(2).start()
+		vs = VideoStream(1).start()
 	# otherwise, grab a reference to the video file
 	else:
 		vs = cv2.VideoCapture(args["video"])
