+++ compvis/Prev Files/visualize_tracking.py
